gdmn-er-bridge: use double precision instead float (#55)

+++ packages/gdmn-er-bridge/src/ddl/builder/DomainResolver.ts
@@ -66,7 +66,7 @@ export class DomainResolver {
       case "TimeStamp":
         return `TIMESTAMP`;
       case "Float":
-        return `FLOAT`;
+        return `DOUBLE PRECISION`;
       case "Boolean":
         return `SMALLINT`;
       case "Blob":
